genie-teams-app/requirements.txt
azure-functions==1.18.0
botbuilder-core==4.16.0
botbuilder-schema==4.16.0
botframework-connector==4.16.0
aiohttp==3.9.3
msrest==0.7.1
PyJWT[crypto]==2.10.1
cryptography>=46.0.0
